Add Markdown syntax
This .tmLanguage file is based on what I was building for MarkDo.
I've spunt it out and plan to support MMD, Pandoc & GFM, but in a
rather loose fashion. The point of this is to be benefical to writers,
esp. when they are writing informally, as in notes, fiction, etc.